calculation-server/requirements.txt
fastapi
uvicorn
geopandas
osrm-py
requests